Starts porting Cogl conformance tests from Clutter
This makes a start on porting the Cogl conformance tests that use to
live in the Clutter repository to be standalone Cogl tests that no
longer require a ClutterStage.

The main thing is that this commit brings in the basic testing
infrastructure we need, so no we can port more and more tests
incrementally.