pipeline: don't prune user pipelines
We encourage developers to create template pipelines that we can
associated re-usable state such as shaders with, but our logic for
pruning redundant ancestry was actually pruning those templates!

We now keep track of which pipelines are created for internal purposes
(e.g. copy-on-write pipelines created when modifying an immutable
pipeline referenced in the journal) and we only ever prune those. We no
longer prune user created pipelines.

We are careful to consider that pipelines that are initially created
explicitly by the user, might eventually become internal only. This
is true if the ref_count of a pipeline is equal to the number of
children it has, since we know the application no longer owns a
reference to the pipeline. In this situation we will allow pruning.

NB: The main reason we need to do pruning is to consider cases where
the user toggles back and fourth between the same states and we need to
avoid creating an endlessly growing chain