Adds ColorMask support to Cogl
This adds CoglPipeline and CoglFramebuffer support for setting a color
mask which is a bit mask defining which color channels should be written
to the current framebuffer.

The framebuffer mask has highest precedence and can be used to affect
all rendering to the framebuffer while the pipeline masks can be used to
affect individual primitives.