Update VS property sheets on enumerations
It seems that the VS projects need to handle the generation of
cogl-enum-types.h and cogl-enum-types.c by themselves again as
cogl-win32-renderer.h needs to be factored in by the glib-mkenums
script on Windows (which would not be done on Linux, where release
tarballs are done).

Discontinue the distribution of cogl-enum-types.c and cogl-enum-types.h
as a result.