mingw-fetch-dependencies.sh: Don't pass -c to wget
The -c option for wget and -C - option to curl are used to make it
continue the download if the file already exists. The idea was that it
wouldn't waste time downloading the files again if the file already
exists. However this causes problems if the remote file gets larger
because the download will continue from the size of the old file so it
will get corrupt. Instead let's just explicitly check if the file
already exists and avoid calling wget or curl altogether.

Reviewed-by: Robert Bragg <robert@linux.intel.com>

(cherry picked from commit c2721664aac000588e0a20b0bb1f8e896b18eddc)