journal: Resort into non-overlapping batches with the same pipeline
Rather than having a single linear list of quads in the journal, it is
now split into an array of batches. Each batch contains a list of
entries which all use the same pipeline. The batches are rendered in
the order they are stored in the array. Each batch also has a bounding
rectangle. When a new rectangle is added to the journal it will try to
find a batch with a matching pipeline by walking backwards through the
array of batches from the end. If the pipeline for the new quad
doesn't match the last batch it can attempt to add it to the batch
behind the last batch as long as the quad doesn't intersect the
batch's bounding box.

The batch's bounding boxes are stored in screen co-ordinates. For this
to work each quad must be fully projected into screen space when it is
added. To avoid calculating the modelview transformed vertices twice
they are now stored alongside the journal entry.

The journal entries are now separately slice allocated and linked in a
list rather than being stored in an array. The position and vertex
data for the entry is now stored directly in the entry struct instead
of having a separate array.