cogl-journal: Use a loose region instead of a bounding box
Instead of just recording the bounding box of a journal batch, it now
uses a new data type called a loose region. A loose region keeps track
of up to 4 separate bounding boxes to represent at least the region
covered by the batch. If a new rectangle added to the region only
increases the size of an existing bounding box to within a certain
threshold then it will be added to that bounding box, otherwise it
will start a new bounding box. This increases the chances that batches
can be grouped together because it can represent sparse regions to a
degree without as much of the complexity as totally accurately
tracking the region.