stash: virtual-framebuffer work in progress
TODO:

would it be worth caching the glScissor associated with a framebuffer so
that when we switch between framebuffers (for virtual framebuffers) we
can cheaply restore the correct glScissor so we don't need to hit the
clip_stack_flush code.

If we know the stencil buffer matches the current clip stack, we know
any clip planes remain constant

as we rotate between framebuffers, we are going to want to set a
transient offset viewport but don't want that to mark the clip state as
changed

when drawing a primitive on multiple framebuffers we should always check
what the current scissor is to see if we can trivially avoid drawing to
multiple framebuffers.