Allow the winsys backend to force rendering onscreen buffers flipped
This adds the option for a winsys to report which kinds of buffers
need a flipped rendering. This will be used by the KMS backend because
that also needs flipped rendering for onscreen buffers. There is a
virtual winsys function to report a bitmask of types of buffers that
need to be flipped. The default implementation just reports offscreen
buffers. None of the winsyss actually implement it yet.

When the backend flips for both types of framebuffer or neither type
of framebuffer the vertend now hardcodes the vector for the flip
instead of adding the uniform.