examples: Adds a more advanced hairy cube example
This example uses the Shell Texturing technique to extrude hair from an
arbitrary mesh (A cube in this example). It doesn't currently add fins
on the edges so you can see shimmering when the shells become parallel
to the viewing vector but nevertheless its quite a nice demonstration of
using Cogl for something more than just basic textured rectangles.

This example also uses the new glsl snippets api and shows how easy it
is to customize the pipeline with your own glsl but at the same time
build on code that's auto generated by Cogl.

Reviewed-by: Neil Roberts <neil@linux.intel.com>