onscreen: Adds cogl_onscreen_start_frame() api
This adds cogl_onscreen_start_frame() api to allow applications to
explicitly start a frame without rendering anything. This will ensure
that all ancillary buffers associated with the onscreen framebuffer have
been allocated and the framebuffer is prepared for rendering. This also
gives applications an opportunity to query properties of the framebuffer
that may be affected by starting a new frame that might also be
useful to know before anything is actually drawn; such as the width
and height of the framebuffer.

The plan is to follow up with another patch that lets applications query
the age of the back buffer contents which will allow applications to
efficiently perform incremental updates of reused buffers, and this
relies on us being able to start the next frame without rendering
anything.