onscreen: Adds buffer age api
This adds a cogl_oncsreen_get_back_buffer_age() function that allows
applications to query how old the back buffer contents are at the start
of a new frame.

The intention of this api is to help optimize applications that
frequently only update small regions of the screen.  Applications can
maintain a running record of the regions changed over the last three or
so frames and at the start of each frame they can check the age of the
back buffer contents and if its between 1 and 3 then they can use their
history of the last three frames to calculate how to repair the contents
of the re-used buffer instead of having to redraw everything.

A nice advantage this approach has is that it doesn't require a blit to
present the incremental updates to the front buffer, a flip can be used
which is more efficient.