texture-pixmap-x11.h: define __COGL_XLIB_H_INSIDE__
Ideally we wouldn't have exposed cogl-texture-pixmap-x11.h as a
top level header and would have just automatically included it in cogl.h
but we already have code that assumes it can be directly included.

This ensures we define __COGL_XLIB_H_INSIDE__ as a reminder that
it is a top level header in case we later need to include other
cogl internal cogl headers which would cause a build time error without
this defined.