cogl-hair: slant shells to avoid banding
As an experiment this patch adds a varying slant to all of the shells so
that they are not parallel to avoid them becoming invisible when they
are parallel with the viewing direction.

This still needs some further work though because the shell textures
need to be updated according to the sloping too otherwise they don't
appear to extrude the hair continuously from a single point.