bindings: Starts to add v8 JavaScript bindings
This starts to add support for v8 JavaScript bindings for Cogl.

Since JavaScript bindings only make sense as an embedded language this
patch adds a minimal cogl-v8 program that is used to bootstrap an
embedding environment that is initially also responsible for creating
a CoglContext and a single CoglOnscreen framebuffer, this will no
doubt change as the bindings evolve.

Eventually we may provide a cogl-v8 sub-library that might have a public
api like cogl_v8_bind (v8::Handle<v8::ObjectTemplate> global) that
can add the bindings to an object and context created by something else
(such as node.js for example).

At this point the api coverage is basically zero, but the aim next is
to look at auto generating bindings for most of Cogl's trivial methods.