cogl-sharp: Start looking at exposing Context, FrameBuffer and OnScreen
This is by not mean complete but a first stab at looking at what is
needed to bind Cogl to C#.

parse-gir.py is expected to grow a bit to generate the bulk of the
bindings, for now it was used to generate the BufferBit bitfield.

Life cycle of objects is not looked at in details yet.