Allow lazy texture storage allocation
Consistent with how we lazily allocate framebuffers this patch allows us
to instantiate textures but still specify constraints and requirements
before allocating storage so that we can be sure to allocate the most
appropriate/efficient storage.

This adds a cogl_texture_allocate() function that is analogous to
cogl_framebuffer_allocate() which can optionally be called to gracefully
catch any errors with texture allocation but if this function isn't used
explicitly then Cogl will implicitly ensure textures are allocated
before the storage is needed.