rename CoglFrameTimings to CoglSwapInfo
The aim is to pass an opaque structure to the swap buffers notification
callbacks and enable access to frame timing information this way. Since
there could be numerous non-timing related things to report here too
this gives the structure a more general name.

TODO: fold this back into owen's patch